+++ tacs/pytacs.py
@@ -37,7 +37,9 @@ import numpy as np
 from mpi4py import MPI
 import warnings
 import tacs.TACS, tacs.constitutive, tacs.elements, tacs.functions, tacs.problems.static
-from pyNastran.bdf.bdf import read_bdf
+from tacs.pymeshloader import pyMeshLoader
+
+DEG2RAD = np.pi / 180.0
 
 warnings.simplefilter('default')
 try:
@@ -49,9 +51,6 @@ except ImportError:
         print("Could not find any OrderedDict class. "
               "For python 2.6 and earlier, use:"
               "\n pip install ordereddict")
-
-DEG2RAD = numpy.pi / 180.0
-
 
 class pyTACS(object):
 
@@ -121,6 +120,7 @@ class pyTACS(object):
             'numberSolutions': [bool, True],
             'printTiming': [bool, False],
             'printIterations': [bool, True],
+            'printDebug': [bool, False],
 
         }
 
@@ -151,14 +151,23 @@ class pyTACS(object):
         importTime = time.time()
 
         # Create and load mesh loader object.
-        self._scanBdfFile(fileName)
+        debugFlag = self.getOption('printDebug')
+        self.meshLoader = pyMeshLoader(self.comm, self.dtype, debugFlag)
+        self.meshLoader.scanBdfFile(fileName)
         self.bdfName = fileName
+        # Save pynastran bdf object
+        self.bdfInfo = self.meshLoader.getBDFInfo()
 
         meshLoadTime = time.time()
 
         # Retrieve the number of components. This is the maximum
-        # number of unique constituive objects possible in this model.
-        self.nComp = self.bdfInfo.nproperties
+        # number of unique constitutive objects possible in this model.
+        self.nComp = self.meshLoader.getNumComponents()
+
+        # Load all the component descriptions
+        self.compDescripts = self.meshLoader.getComponentDescripts()
+        self.elemDescripts = self.meshLoader.getElementDescripts()
+
 
         # Set the starting dvNum and scaleList
         self.dvNum = dvNum
@@ -186,9 +195,6 @@ class pyTACS(object):
         self.dvSensList = OrderedDict()
         self.xptSensList = OrderedDict()
 
-        # Element IDs (tacs ordering) partitioned to this proc
-        self.globalToLocalNodeIDDict = None
-
         # List of initial coordinates
         self.coords0 = None
 
@@ -202,8 +208,7 @@ class pyTACS(object):
         # Flag for mat/vector creation
         self._variablesCreated = False
 
-        # TACS assembler and creator objects
-        self.creator = None
+        # TACS assembler object
         self.assembler = None
 
         initFinishTime = time.time()
@@ -219,567 +224,6 @@ class pyTACS(object):
             self.pp('|')
             self.pp('| %-30s: %10.3f sec' % ('TACS Total Initialization Time', initFinishTime - startTime))
             self.pp('+--------------------------------------------------+')
-
-    ####### pyNastran BDF parsing methods ########
-
-    def _scanBdfFile(self, fileName):
-        """
-        Scan nastran bdf file using pyNastran's bdf parser.
-        We also set up arrays that will be require later to build tacs.
-        """
-
-        # Only print debug info on root
-        if self.comm.rank == 0:
-            debugPrint = True
-        else:
-            debugPrint = False
-
-        # Read in bdf file as pynsatran object
-        # By default we avoid cross-referencing unless we actually need it,
-        # since its expensive for large models
-        self.bdfInfo = read_bdf(fileName, validate=False, xref=False, debug=debugPrint)
-        # Set flag letting us know model is not xrefed yet
-        self.bdfInfo.is_xrefed = False
-        self.bdfInfo.no_properties = False
-
-        # If no property cards were found in bdf, we have to add dummy cards
-        # so pynastran doesn't through errors when cross-referencing
-        if self.bdfInfo.nproperties == 0:
-            # Flag to warn the user that the property cards are placeholders
-            # and should not be read in using pytacs _elemCallBackFromBDF method
-            self.bdfInfo.no_properties = True
-            # If no material properties were found,
-            # add dummy properties and materials
-            matID = 1
-            E = 70.0
-            G = 35.0
-            nu = 0.3
-            self.bdfInfo.add_mat1(matID, E, G, nu)
-
-            # Loop through all elements and add dummy property, as necessary
-            for element_id in self.bdfInfo.elements:
-                element = self.bdfInfo.elements[element_id]
-                if element.pid not in self.bdfInfo.property_ids:
-                    self.bdfInfo.add_pbar(element.pid, matID)
-
-        # Create dictionaries for mapping between tacs and nastran id numbering
-        self._updateNastranToTACSDicts()
-
-        # Try to get the node x,y,z locations from bdf file
-        try:
-            self.bdfXpts = self.bdfInfo.get_xyz_in_coord()
-        # If this fails, the file may reference multiple coordinate systems
-        # and will have to be cross-refernced to work
-        except:
-            self.bdfInfo.cross_reference()
-            self.bdfInfo.is_xrefed = True
-            self.bdfXpts = self.bdfInfo.get_xyz_in_coord()
-
-        # element card contained within each property group (may contain multiple per group)
-        # Each entry will eventually have its own tacs element object assigned to it
-        # (ex. self.elemDescripts = [['CQUAD4', 'CTRIA3'], ['CQUAD9', 'CQUAD4'], ['CBAR']])
-        self.elemDescripts = []
-        # Pointer index used to map elemDescript entries into flat list
-        # (ex. self.elementObjectNumByComp = [[0, 1], [2, 3], [4]])
-        self.elemObjectNumByComp = []
-        # User-defined name for property/component group,
-        # this may be read in through ICEMCFD/FEMAP format comments in BDF
-        self.compDescript = []
-
-        # Populate list entries with default values
-        for pID in self.bdfInfo.property_ids:
-            self.elemDescripts.append([])
-            self.elemObjectNumByComp.append([])
-            # Check if there is a Femap label for this component
-            propComment = self.bdfInfo.properties[pID].comment
-            if '$ Femap Property' in propComment:
-                # Pick off last word from comment, this is the name
-                propName = propComment.split()[-1]
-                self.compDescript.append(propName)
-            #  Default component name
-            else:
-                self.compDescript.append('untitled')
-
-        # Element connectivity information
-        self.elemConnectivity = [None] * self.bdfInfo.nelements
-        self.elemConnectivityPointer = [None] * (self.bdfInfo.nelements + 1)
-        self.elemConnectivityPointer[0] = 0
-        elementTypeCounter = 0
-        # List specifying which tacs element object each element in bdf should point to
-        self.elemObjectNumByElem = [None] * (self.bdfInfo.nelements)
-
-        # Loop through every element and record information needed for tacs
-        for tacsElementID, nastranElementID in enumerate(self.bdfInfo.element_ids):
-            element = self.bdfInfo.elements[nastranElementID]
-            elementType = element.type.upper()
-            propertyID = element.pid
-            componentID = self.idMap(propertyID, self.nastranToTACSCompIDDict)
-
-            # This element type has not been added to the list for the component group yet, so we append it
-            if elementType not in self.elemDescripts[componentID]:
-                self.elemDescripts[componentID].append(elementType)
-                self.elemObjectNumByComp[componentID].append(elementTypeCounter)
-                elementTypeCounter += 1
-
-            # Find the index number corresponding to the element object number for this component
-            componentTypeIndex = self.elemDescripts[componentID].index(elementType)
-            self.elemObjectNumByElem[tacsElementID] = self.elemObjectNumByComp[componentID][componentTypeIndex]
-
-            # We've identified a ICEM property label
-            if 'Shell element data for family' in element.comment:
-                componentName = element.comment.split()[-1]
-                self.compDescript[componentID] = componentName
-
-            conn = element.nodes.copy()
-
-            # TACS has a different node ordering than Nastran for certain elements,
-            # we now perform the reordering (if necessary)
-            if elementType in ['CQUAD4', 'CQUADR']:
-                conn = [conn[0], conn[1], conn[3], conn[2]]
-            elif elementType in ['CQUAD9', 'CQUAD']:
-                conn = [conn[0], conn[4], conn[1], conn[7], conn[8], conn[5], conn[3], conn[6], conn[2]]
-            elif elementType in ['CHEXA8', 'CHEXA']:
-                conn = [conn[0], conn[1], conn[3], conn[2], conn[4], conn[5], conn[7], conn[6]]
-
-            # Map node ids in connectivity from Nastan numbering to TACS numbering
-            self.elemConnectivity[tacsElementID] = self.idMap(conn, self.nastranToTACSNodeIDDict)
-            self.elemConnectivityPointer[tacsElementID + 1] = self.elemConnectivityPointer[
-                                                                     tacsElementID] + len(element.nodes)
-
-        # Allocate list for user-specified tacs element objects
-        self.elemObjects = [None] * elementTypeCounter
-
-    def _updateNastranToTACSDicts(self):
-        '''
-        Create dictionaries responsible for mapping over
-        global node, global element, and property/component ID numbers
-        from NASTRAN (as read in from BDF) to TACS (contiguous, 0-based).
-        The keys of each dictionary are the NASTRAN ID and the entry the TACS ID,
-        such that:
-            tacsNodeID = self.nastranToTACSNodeIDDict[nastranNodeID]
-            tacsComponentID = self.nastranToTACSCompIDDict[nastranPropertyID]
-            tacsElementID = self.nastranToTACSNodeIDDict[nastranElementID]
-        The dictionaries contain all elements/nodes found in the BDF,
-        not just those *owned* by this processor
-        '''
-        # Create Node ID map
-        nastranIDs = self.bdfInfo.node_ids
-        tacsIDs = range(self.bdfInfo.nnodes)
-        nodeTuple = zip(nastranIDs, tacsIDs)
-        self.nastranToTACSNodeIDDict = dict(nodeTuple)
-
-        # Create Property/Component ID map
-        nastranIDs = self.bdfInfo.property_ids
-        tacsIDs = range(self.bdfInfo.nproperties)
-        propTuple = zip(nastranIDs, tacsIDs)
-        self.nastranToTACSCompIDDict = dict(propTuple)
-
-        # Create Element ID map
-        nastranIDs = self.bdfInfo.element_ids
-        tacsIDs = range(self.bdfInfo.nelements)
-        elemTuple = zip(nastranIDs, tacsIDs)
-        self.nastranToTACSElemIDDict = dict(elemTuple)
-
-    def _elemCallBackFromBDF(self):
-        """
-        Automatically setup elemCallBack using information contained in BDF file.
-        This function assumes all material properties are specified in the BDF.
-        """
-        # Make sure cross-referencing is turned on in pynastran
-        if self.bdfInfo.is_xrefed is False:
-            self.bdfInfo.cross_reference()
-            self.bdfInfo.is_xrefed = True
-
-        # Check if any properties are in the BDF
-        if self.bdfInfo.no_properties:
-            raise Error("BDF file '%s' has no properties included in it. "
-                        "User must define own elemCallBack function." % (self.bdfName))
-
-        # Create a dictionary to sort all elements by property number
-        elemDict = {}
-        for elementID in self.bdfInfo.elements:
-            element = self.bdfInfo.elements[elementID]
-            propertyID = element.pid
-            if propertyID not in elemDict:
-                elemDict[propertyID] = {}
-                elemDict[propertyID]['elements'] = []
-                elemDict[propertyID]['dvs'] = {}
-            elemDict[propertyID]['elements'].append(element)
-
-        # Create a dictionary to sort all design variables
-        for dv in self.bdfInfo.dvprels:
-            propertyID = self.bdfInfo.dvprels[dv].pid
-            dvName = self.bdfInfo.dvprels[dv].pname_fid
-            self.dvNum = max(self.dvNum, self.bdfInfo.dvprels[dv].dvids[0])
-            elemDict[propertyID]['dvs'][dvName] = self.bdfInfo.dvprels[dv]
-        # Create option for user to specify scale values in BDF
-        self.scaleList = [1.0] * self.dvNum
-
-        # Callback function to return appropriate tacs MaterialProperties object
-        # For a pynastran mat card
-        def matCallBack(matInfo):
-            # First we define the material property object
-            if matInfo.type == 'MAT1':
-                mat = tacs.constitutive.MaterialProperties(rho=matInfo.rho, E=matInfo.e,
-                                                            nu=matInfo.nu, ys=matInfo.St,
-                                                            alpha=matInfo.a)
-            elif matInfo.type == 'MAT8':
-                E1 = matInfo.e11
-                E2 = matInfo.e22
-                nu12 = matInfo.nu12
-                G12 = matInfo.g12
-                G13 = matInfo.g1z
-                G23 = matInfo.g2z
-                # If out-of-plane shear values are 0, Nastran defaults them to the in-plane
-                if G13 == 0.0:
-                    G13 = G12
-                if G23 == 0.0:
-                    G23 = G12
-                rho = matInfo.rho
-                Xt = matInfo.Xt
-                Xc = matInfo.Xc
-                Yt = matInfo.Yt
-                Yc = matInfo.Yc
-                S12 = matInfo.S
-                # TODO: add alpha
-                mat = tacs.constitutive.MaterialProperties(rho=rho, E1=E1, E2=E2, nu12=nu12, G12=G12, G13=G13, G23=G23,
-                                                            Xt=Xt, Xc=Xc, Yt=Yt, Yc=Yc, S12=S12)
-            else:
-                raise Error("Unsupported material type '%s' for material number %d. " % (matInfo.type, matInfo.mid))
-            
-            return mat
-
-        def elemCallBack(dvNum, compID, compDescript, elemDescripts, globalDVs, **kwargs):
-            # Initialize scale list for design variables we will add
-            scaleList = []
-            
-            # Get the Nastran property ID
-            propertyID = kwargs['propID']
-            propInfo = self.bdfInfo.properties[propertyID]
-            elemInfo = elemDict[propertyID]['elements'][0]
-            
-            # First we define the material object
-            # This property only references one material
-            if hasattr(propInfo, 'mid_ref'):
-                matInfo = propInfo.mid_ref
-                mat = matCallBack(matInfo)
-            # This property references multiple materials (maybe a laminate)
-            elif hasattr(propInfo, 'mids_ref'):
-                mat = []
-                for matInfo in propInfo.mids_ref:
-                    mat.append(matCallBack(matInfo))
-
-            # Next we define the constitutive object
-            if propInfo.type == 'PSHELL':  # Nastran isotropic shell
-                kcorr = propInfo.tst
-
-                if 'T' in elemDict[propertyID]['dvs']:
-                    thickness = elemDict[propertyID]['dvs']['T'].dvids_ref[0].xinit
-                    tNum = elemDict[propertyID]['dvs']['T'].dvids[0] - 1
-                    minThickness = elemDict[propertyID]['dvs']['T'].dvids_ref[0].xlb
-                    maxThickness = elemDict[propertyID]['dvs']['T'].dvids_ref[0].xub
-                    name = elemDict[propertyID]['dvs']['T'].dvids_ref[0].label
-                    self.scaleList[tNum - 1] = elemDict[propertyID]['dvs']['T'].coeffs[0]
-                else:
-                    thickness = propInfo.t
-                    tNum = -1
-                    minThickness = 0.0
-                    maxThickness = 1e20
-
-                con = tacs.constitutive.IsoShellConstitutive(mat, t=thickness,
-                                                             tlb=minThickness, tub=maxThickness, tNum=tNum)
-
-            elif propInfo.type == 'PCOMP':  # Nastran composite shell
-                numPlies = propInfo.nplies
-                plyThicknesses = []
-                plyAngles = []
-                plyMats = []
-
-                # if the laminate is symmetric, mirror the ply indices
-                if propInfo.lam == 'SYM':
-                    plyIndices = range(numPlies / 2)
-                    plyIndices.extend(plyIndices[::-1])
-                else:
-                    plyIndices = range(numPlies)
-
-                # Loop through plies and setup each entry in layup
-                for ply_i in plyIndices:
-                    plyThicknesses.append(propInfo.thicknesses[ply_i])
-                    plyMat = tacs.constitutive.OrthotropicPly(plyThicknesses[ply_i], mat[ply_i])
-                    plyMats.append(plyMat)
-                    plyAngles.append(propInfo.thetas[ply_i] * DEG2RAD)
-
-                # Convert thickness/angles to appropriate numpy array
-                plyThicknesses = np.array(plyThicknesses, dtype=self.dtype)
-                plyAngles = np.array(plyAngles, dtype=self.dtype)
-
-                if propInfo.lam is None or propInfo.lam in ['SYM', 'MEM']:
-                    # Discrete laminate class (not for optimization)
-                    con = tacs.constitutive.CompositeShellConstitutive(plyMats, plyThicknesses, plyAngles)
-                    # Need to add functionality to consider only membrane in TACS for type = MEM
-
-                else:
-                    raise Error("Unrecognized LAM type '%s' for PCOMP number %d." % (propInfo.lam, propertyID))
-
-            elif propInfo.type == 'PSOLID':  # Nastran solid property
-                if 'T' in elemDict[propertyID]['dvs']:
-                    thickness = elemDict[propertyID]['dvs']['T'].dvids_ref[0].xinit
-                    tNum = elemDict[propertyID]['dvs']['T'].dvids[0] - 1
-                    minThickness = elemDict[propertyID]['dvs']['T'].dvids_ref[0].xlb
-                    maxThickness = elemDict[propertyID]['dvs']['T'].dvids_ref[0].xub
-                    name = elemDict[propertyID]['dvs']['T'].dvids_ref[0].label
-                    self.scaleList[tNum - 1] = elemDict[propertyID]['dvs']['T'].coeffs[0]
-                else:
-                    thickness = 1.0
-                    tNum = -1
-                    minThickness = 0.0
-                    maxThickness = 10.0
-
-                con = tacs.constitutive.SolidConstitutive(mat, t=thickness,
-                                                          tlb=minThickness, tub=maxThickness, tNum=tNum)
-
-            else:
-                raise Error("Unsupported property type '%s' for property number %d. " % (propInfo.type, propertyID))
-
-            # Set up transform object which may be required for certain elements
-            transform = None
-            if hasattr(elemInfo, 'theta_mcid_ref'):
-                mcid = elemDict[propertyID]['elements'][0].theta_mcid_ref
-                if mcid:
-                    if mcid.type == 'CORD2R':
-                        refAxis = mcid.i
-                        transform = tacs.elements.ShellRefAxisTransform(refAxis)
-                    else:  # Don't support spherical/cylindrical yet
-                        raise Error("Unsupported material coordinate system type "
-                                    "'%s' for property number %d." % (mcid.type, propertyID))
-
-            # Finally set up the element objects belonging to this component
-            elemList = []
-            for descript in elemDescripts:
-                if descript in ['CQUAD4', 'CQUADR']:
-                    elem = tacs.elements.Quad4Shell(transform, con)
-                elif descript in ['CQUAD9', 'CQUAD']:
-                    elem = tacs.elements.Quad9Shell(transform, con)
-                elif descript in ['CTRIA3', 'CTRIAR']:
-                    elem = tacs.elements.Tri3Shell(transform, con)
-                elif 'CTETRA' in descript:
-                    # May have variable number of nodes in card
-                    nnodes = len(elemInfo.nodes)
-                    if nnodes == 4:
-                        basis = tacs.elements.LinearTetrahedralBasis()
-                    elif nnodes == 10:
-                        basis = tacs.elements.QuadraticTetrahedralBasis()
-                    else:
-                        raise Error("TACS does not currently support CTETRA elements with %d nodes."%nnodes)
-                    model = tacs.elements.LinearElasticity3D(con)
-                    elem = tacs.elements.Element3D(model, basis)
-                elif descript in ['CHEXA8', 'CHEXA']:
-                    basis = tacs.elements.LinearHexaBasis()
-                    model = tacs.elements.LinearElasticity3D(con)
-                    elem = tacs.elements.Element3D(model, basis)
-                else:
-                    raise Error("Unsupported element type "
-                                "'%s' specified for property number %d." % (descript, propertyID))
-                elemList.append(elem)
-
-            return elemList, scaleList
-
-        return elemCallBack
-
-    def getBDFInfo(self):
-        """
-        Return pynastran bdf object.
-        """
-        return self.bdfInfo
-
-    def getNumBDFNodes(self):
-        '''
-        Return number of nodes found in bdf.
-        '''
-        return self.bdfInfo.nnodes
-
-    def getNumOwnedNodes(self):
-        '''
-        Return number of nodes owned by this processor.
-        '''
-        return self.assembler.getNumOwnedNodes()
-
-    def getNumBDFElements(self):
-        '''
-        Return number of elements found in bdf.
-        '''
-        return self.bdfInfo.nelements
-
-    def getBDFNodes(self, nodeIDs, nastranOrdering=False):
-        '''
-        Return x,y,z location of specified node in bdf file.
-        '''
-        # Convert to tacs numbering, if necessary
-        if nastranOrdering:
-            nodeIDs = self.idMap(nodeIDs, self.nastranToTACSNodeIDDict)
-        return self.bdfXpts[nodeIDs]
-
-    def getElementComponents(self):
-        elements = self.bdfInfo.elements
-        propertyIDList = [elements[eID].pid for eID in self.bdfInfo.element_ids]
-        compIDList = self.idMap(propertyIDList, self.nastranToTACSCompIDDict)
-        return compIDList
-
-    def getConnectivityForComp(self, componentID, nastranOrdering=False):
-        # Find all of the element IDs belonging to this property group
-        propertyID = list(self.bdfInfo.property_ids)[componentID]
-        elementIDs = self.bdfInfo.get_element_ids_dict_with_pids()[propertyID + 1]
-        compConn = []
-        for elementID in elementIDs:
-            # We've now got the connectivity for this element, but it is in nastrans node numbering
-            nastranConn = self.bdfInfo.elements[elementID].nodes
-            if nastranOrdering:
-                compConn.append(nastranConn)
-            else:
-                # Convert Nastran node numbering back to tacs numbering
-                tacsConn = [self.bdfInfo.nid_map[nID] for nID in nastranConn]
-                # Append element connectivity to list for component
-                compConn.append(tacsConn)
-        return compConn
-
-    def getElementObjectNumsForComp(self, componentID):
-        """
-        Return tacs element object number for each element type
-        belonging to this component.
-        """
-        return self.elemObjectNumByComp[componentID][:]
-
-    def getLocalNodeIDsFromGlobal(self, globalIDs, nastranOrdering=False):
-        """
-        given a list of node IDs in global (non-partitioned) ordering
-        returns the local (partitioned) node IDs on each processor.
-        If a requested node is not included on this processor,
-        an entry of -1 will be returned.
-        """
-        # Convert to tacs node numbering, if necessary
-        if nastranOrdering:
-            globalIDs = self.idMap(globalIDs, self.nastranToTACSNodeIDDict)
-
-        # Ensure input is list-like
-        globalIDs = np.atleast_1d(globalIDs)
-
-        # Get the node id offset for this processor
-        OwnerRange = self.assembler.getOwnerRange()
-        nodeOffset = OwnerRange[self.comm.rank]
-
-        # Get the local ID numbers for this proc
-        tacsLocalIDs = []
-        for gID in globalIDs:
-            lIDs = self.creator.getAssemblerNodeNums(self.assembler,
-                                                     np.array([gID], dtype=np.intc))
-            # Node was not found on this proc, return -1
-            if len(lIDs) == 0:
-                tacsLocalIDs.append(-1)
-            # Node was found on this proc, shift by nodeOffset to get local index for node
-            else:
-                tacsLocalIDs.append(lIDs[0] - nodeOffset)
-
-        return tacsLocalIDs
-
-    def getLocalElementIDsFromGlobal(self, globalIDs, nastranOrdering=False):
-        """
-        given a list of element IDs in global (non-partitioned) ordering
-        returns the local (partitioned) element IDs on each processor.
-        If a requested element is not included on this processor,
-        an entry of -1 will be returned.
-        """
-        # Convert to tacs node numbering, if necessary
-        if nastranOrdering:
-            globalIDs = self.idMap(globalIDs, self.nastranToTACSElemIDDict)
-
-        # Ensure input is list-like
-        globalIDs = np.atleast_1d(globalIDs)
-
-        # Get the node id offset for this processor
-        OwnerRange = self.assembler.getOwnerRange()
-        nodeOffset = OwnerRange[self.comm.rank]
-
-        # Get the local ID numbers for this proc
-        tacsLocalIDs = []
-        for gID in globalIDs:
-            # element was found on this proc, get local ID num
-            if gID in self.globalToLocalNodeIDDict:
-                lID = self.globalToLocalNodeIDDict[gID]
-            # element was not found on this proc, return -1
-            else:
-                lID = -1
-            tacsLocalIDs.append(lID)
-
-        return tacsLocalIDs
-
-    def getGlobalElementIDsForComps(self, componentIDs, nastranOrdering=False):
-        """
-        Returns a list of element IDs belonging to specified components
-        """
-        # Make sure list is flat
-        componentIDs = self._flatten(componentIDs)
-        # Convert tacs component IDs to nastran property IDs
-        propertyIDs = [0] * len(componentIDs)
-        for i, componentID in enumerate(componentIDs):
-            propertyIDs[i] = list(self.bdfInfo.property_ids)[componentID]
-        # Get dictionary whose values are the element ids we are looking for
-        elementIDDict = self.bdfInfo.get_element_ids_dict_with_pids(propertyIDs)
-        # Convert to list
-        elementIDs = list(elementIDDict.values())
-        # Make sure list is flat
-        elementIDs = self._flatten(elementIDs)
-        # Convert to tacs element numbering, if necessary
-        if not nastranOrdering:
-            elementIDs = self.idMap(elementIDs, self.nastranToTACSElemIDDict)
-        return elementIDs
-
-    def getLocalElementIDsForComps(self, componentIDs):
-        """
-        Get the local element numbers on each proc used by tacs
-        corresponding to the component groups in componentIDs.
-        """
-        if self.creator is None:
-            raise Error("TACS assembler has not been created. "
-                        "Assembler must created first by running 'createTACS' method.")
-        # Make sure list is flat
-        componentIDs = self._flatten(componentIDs)
-        # Get the element object IDs belonging to each of these components
-        objIDs = []
-        for componentID in componentIDs:
-            tmp = self.getElementObjectNumsForComp(componentID)
-            objIDs.extend(tmp)
-        objIDs = np.array(objIDs, dtype=np.intc)
-        # Get the local element IDs corresponding to the object IDs on this processor (if any)
-        localElemIDs = self.creator.getElementIdNums(objIDs)
-        return list(localElemIDs)
-
-    def _getGlobalToLocalNodeIDDict(self):
-        """
-        Creates a dictionary who's keys correspond to the global ID of each node (tacs ordering)
-        owned by this processor and whose values correspond to the local node index for this proc.
-        The dictionary can be used to convert from a global node ID to local using the assignment below*:
-        localNodeID = globalToLocalNodeIDDict[globalNodeID]
-
-        * assuming globalNodeID is owned on this processor
-        """
-        # Do sorting on root proc
-        if self.comm.rank == 0:
-            # List containing which poc every element belongs to
-            elemPartition = self.creator.getElementPartition()
-            # Create an empty list that we will use to sort what elements are on what procs
-            allOwnedElementIDs = [[] for proc_i in range(self.comm.size)]
-            for elemGlobalID in range(len(elemPartition)):
-                ownerProc = elemPartition[elemGlobalID]
-                allOwnedElementIDs[ownerProc].append(elemGlobalID)
-        else:
-            allOwnedElementIDs = None
-
-        # Scatter the list from the root so each proc knows what element ID it owns
-        ownedElementIDs = self.comm.scatter(allOwnedElementIDs, root=0)
-        # Create dictionary that gives the corresponding local ID for each global ID owned by this proc
-        globalToLocalNodeIDDict = { gID : lID for lID, gID in enumerate(ownedElementIDs) }
-
-        return globalToLocalNodeIDDict
 
     def addGlobalDV(self, descript, value,
                     lower=None, upper=None, scale=1.0):
@@ -942,10 +386,6 @@ class pyTACS(object):
         if exclude is not None:
             excludeIDs = self._getCompIDs(excludeOp, exclude)
 
-        if includeBounds is not None:
-                includeBoundIDs = \
-                    self._getCompIDsInBounds(includeBounds, projectVector)
-
         iSet = set(includeIDs)
         eSet = set(excludeIDs)
 
@@ -978,7 +418,7 @@ class pyTACS(object):
             # sort them
             compDescript = []
             for i in range(len(compIDs)):
-                compDescript.append(self.compDescript[compIDs[i]])
+                compDescript.append(self.compDescripts[compIDs[i]])
 
             # define a general argsort
             def argsort(seq):
@@ -1045,10 +485,6 @@ class pyTACS(object):
             Argument passed to selctCompIDs. See this function for
             more information
 
-        includeBounds : varries
-            Argument passed to selctCompIDs. See this function for
-            more information
-
         compIDs: list
             List of compIDs to select. Alternative to selectCompIDs
             arguments.
@@ -1065,7 +501,7 @@ class pyTACS(object):
 
         # Flatten and get element numbers on each proc corresponding to specified compIDs
         compIDs = self._flatten(compIDs)
-        elemIDs = self.getLocalElementIDsForComps(compIDs)
+        elemIDs = self.meshLoader.getLocalElementIDsForComps(compIDs)
 
         # We try to setup the function, if it fails it may not be implimented:
         try:
@@ -1105,16 +541,9 @@ class pyTACS(object):
         compIDs = self._flatten(compIDs)
         compDescripts = []
         for i in range(len(compIDs)):
-            compDescripts.append(self.compDescript[compIDs[i]])
+            compDescripts.append(self.compDescripts[compIDs[i]])
 
         return compDescripts
-
-    def getElementDescripts(self):
-        """
-        Get nested list containing all element types owned by each component group
-        example: [['CQUAD4', 'CTRIA3], ['CQUAD4'], ['CQUAD4', CQUAD9']]
-        """
-        return self.elemDescripts
 
     def getFunctionKeys(self):
         """Return a list of the current function key names"""
@@ -1171,189 +600,6 @@ class pyTACS(object):
         # Reset the Aitken acceleration for multidisciplinary analyses
         self.doDamp = False
 
-    def _setupTACSCreator(self):
-        """
-        Setup TACSCreator object responsible for creating TACSAssembler
-        """
-        self.creator = tacs.TACS.Creator(self.comm, self.varsPerNode)
-
-        # Append RBE elements to element list, these are not setup by the user
-        for rbe in self.bdfInfo.rigid_elements.values():
-            if rbe.type == 'RBE2':
-                self._addTACSRBE2(rbe)
-            elif rbe.type == 'RBE3':
-                self._addTACSRBE3(rbe)
-            else:
-                raise NotImplementedError("Rigid element of type '{}' is not supported".format(rbe.type))
-
-        if self.comm.rank == 0:
-            # Set connectivity for all elements
-            ptr = np.array(self.elemConnectivityPointer, dtype=np.intc)
-            conn = np.array(self._flatten(self.elemConnectivity), dtype=np.intc)
-            objectNums = np.array(self.elemObjectNumByElem, dtype=np.intc)
-            self.creator.setGlobalConnectivity(self.bdfInfo.nnodes, ptr, conn, objectNums)
-
-            # Set up the boundary conditions
-            bcDict = {}
-            for spc_id in self.bdfInfo.spcs:
-                for spc in self.bdfInfo.spcs[spc_id]:
-                    # Loop through every node specifed in this spc and record bc info
-                    for j, nastranNode in enumerate(spc.nodes):
-                        tacsNode = self.idMap(nastranNode, self.nastranToTACSNodeIDDict)
-                        # If node hasn't been added to bc dict yet, add it
-                        if tacsNode not in bcDict:
-                            bcDict[tacsNode] = {}
-
-                        # Loop through each dof and record bc info if it is included in this spc
-                        for dof in range(self.varsPerNode):
-                            # Add 1 to get nastran dof number
-                            nastranDOF = dof + 1
-                            if spc.type == 'SPC':
-                                # each node may have its own dofs uniquely constrained
-                                constrainedDOFs = spc.components[j]
-                                # The boundary condition may be forced to a non-zero value
-                                constrainedVal = spc.enforced[j]
-                            else:  # SPC1?
-                                # All nodes always have the same dofs constrained
-                                constrainedDOFs = spc.components
-                                # This boundary condition is always 0
-                                constrainedVal = 0.0
-                            # if nastran dof is in spc components string, add it to the bc dict
-                            if self._isDOFInString(constrainedDOFs, nastranDOF):
-                                bcDict[tacsNode][dof] = constrainedVal
-
-
-            # Convert bc information from dict to list
-            bcnodes = []
-            bcdofs = []
-            bcptr = [0]
-            bcvals = []
-            numbcs = 0
-            for tacsNode in bcDict:
-                bcnodes.append(tacsNode)
-                # Store constrained dofs for this node
-                dofs = bcDict[tacsNode].keys()
-                bcdofs.extend(dofs)
-                # Store enforced bc value
-                vals = bcDict[tacsNode].values()
-                bcvals.extend(vals)
-                # Increment bc pointer with how many constraints weve added for this node
-                numbcs += len(bcDict[tacsNode])
-                bcptr.append(numbcs)
-
-            # Recast lists as numpy arrays
-            bcnodes = np.array(bcnodes, dtype=np.intc)
-            bcdofs = np.array(bcdofs, dtype=np.intc)
-            bcptr = np.array(bcptr, dtype=np.intc)
-            bcvals = np.array(bcvals, dtype=self.dtype)
-            # Set boundary conditions in tacs
-            self.creator.setBoundaryConditions(bcnodes, bcptr, bcdofs, bcvals)
-
-            # Set node locations
-            Xpts = self.bdfInfo.get_xyz_in_coord().astype(tacs.TACS.dtype)
-            self.creator.setNodes(Xpts.flatten())
-
-        # Set the elements for each component
-        self.creator.setElements(self.elemObjects)
-
-    def _isDOFInString(self, constrained_dofs, dof):
-        """
-        Find if dof number (nastran numbering) occurs in constraint string.
-
-        Parameters
-        ----------
-        constrained_dofs : string
-            String containing list of dofs (ex. '123456')
-        dof : int or string
-            nastran dof number to check for
-        """
-        # Convert to string, if necessary
-        if isinstance(dof, int):
-            dof = '%d' % (dof)
-        # pyNastran only supports 0,1,2,3,4,5,6 as valid dof components
-        # For this reason, we'll treat 0 as if its 7, since it's traditionally never used in nastran
-        if dof == '7':
-            dof = '0'
-        location = constrained_dofs.find(dof)
-        # if dof is found, return true
-        if location > -1:
-            return True
-        else:
-            return False
-
-    def _addTACSRBE2(self, rbeInfo):
-        """
-        Method to automatically set up RBE2 element from bdf file for user.
-        User should *NOT* set these up in their elemCallBack function.
-        """
-        indepNode = rbeInfo.independent_nodes
-        depNodes = []
-        depConstrainedDOFs = []
-        dummyNodes = []
-        dofsAsString = rbeInfo.cm
-        dofsAsList = self.isDOFInString(dofsAsString, self.varsPerNode)
-        for node in rbeInfo.dependent_nodes:
-            depNodes.append(node)
-            depConstrainedDOFs.extend(dofsAsList)
-            # add dummy nodes for all lagrange multiplier
-            dummyNodeNum = list(self.bdfInfo.node_ids)[-1] + 1  # Next available node number
-            # Add the dummy node coincident to the dependent node in x,y,z
-            self.bdfInfo.add_grid(dummyNodeNum, self.bdfInfo.nodes[node].xyz)
-            dummyNodes.append(dummyNodeNum)
-
-        conn = indepNode + depNodes + dummyNodes
-        nTotalNodes = len(conn)
-        # Update Nastran to TACS ID mapping dicts, since we just added new nodes to model
-        self._updateNastranToTACSDicts()
-        # Append RBE information to the end of the element lists
-        self.elemConnectivity.append(self.idMap(conn, self.nastranToTACSNodeIDDict))
-        self.elemConnectivityPointer.append(self.elemConnectivityPointer[-1] + nTotalNodes)
-        rbeObj = tacs.elements.RBE2(nTotalNodes, np.array(depConstrainedDOFs, dtype=np.intc))
-        self.elemObjectNumByElem.append(self.elemObjectNumByElem[-1] + 1)
-        self.elemObjects.append(rbeObj)
-        return
-
-    def _addTACSRBE3(self, rbeInfo):
-        """
-        Method to automatically set up RBE3 element from bdf file for user.
-        User should *NOT* set these up in their elemCallBack function.
-        """
-        depNode = rbeInfo.dependent_nodes
-        depConstrainedDOFs = self.isDOFInString(rbeInfo.refc, self.varsPerNode)
-
-        # add dummy node for lagrange multipliers
-        dummyNodeNum = list(self.bdfInfo.node_ids)[-1] + 1  # Next available node number
-        # Add the dummy node coincident to the dependent node in x,y,z
-        self.bdfInfo.add_grid(dummyNodeNum, self.bdfInfo.nodes[depNode[0]].xyz)
-        dummyNodes = [dummyNodeNum]
-        # Update Nastran to TACS ID mapping dicts, since we just added new nodes to model
-        self._updateNastranToTACSDicts()
-
-        indepNodes = []
-        indepWeights = []
-        indepConstrainedDOFs = []
-        for depNodeGroup in rbeInfo.wt_cg_groups:
-            wt = depNodeGroup[0]
-            dofsAsString = depNodeGroup[1]
-            dofsAsList = self.isDOFInString(dofsAsString, self.varsPerNode)
-            for node in depNodeGroup[2]:
-                indepNodes.append(node)
-                indepWeights.append(wt)
-                indepConstrainedDOFs.extend(dofsAsList)
-
-        conn = depNode + indepNodes + dummyNodes
-        nTotalNodes = len(conn)
-        # Append RBE information to the end of the element lists
-        self.elemConnectivity.append(self.idMap(conn, self.nastranToTACSNodeIDDict))
-        self.elemConnectivityPointer.append(self.elemConnectivityPointer[-1] + nTotalNodes)
-        rbeObj = tacs.elements.RBE3(nTotalNodes,
-                                    np.array(depConstrainedDOFs, dtype=np.intc),
-                                    np.array(indepWeights),
-                                    np.array(indepConstrainedDOFs, dtype=np.intc))
-        self.elemObjects.append(rbeObj)
-        self.elemObjectNumByElem.append(self.elemObjectNumByElem[-1] + 1)
-        return
-
     def createTACSAssembler(self, elemCallBack=None):
         """
         This is the 'last' function to be called during the setup. The
@@ -1402,16 +648,221 @@ class pyTACS(object):
         self._createOutputGroups()
         self._createElements(elemCallBack)
 
-        self._setupTACSCreator()
-        self.assembler = self.creator.createTACS()
-
-        self.globalToLocalNodeIDDict = self._getGlobalToLocalNodeIDDict()
+        self.assembler = self.meshLoader.createTACSAssembler(self.varsPerNode)
 
         self._createVariables()
         self._createOutputViewer()
 
-        # Initial set of nodes for geometry manipuation if necessary
+        # Initial set of nodes for geometry manipulation if necessary
         self.coords0 = self.getCoordinates()
+
+    def _elemCallBackFromBDF(self):
+        """
+        Automatically setup elemCallBack using information contained in BDF file.
+        This function assumes all material properties are specified in the BDF.
+        """
+
+        # Check if any properties are in the BDF
+        if self.bdfInfo.missing_properties:
+            raise Error("BDF file '%s' has missing properties cards. "
+                        "Set 'debugPrint' option to True for more information."
+                        "User must define own elemCallBack function." % (self.bdfName))
+
+        # Make sure cross-referencing is turned on in pynastran
+        if self.bdfInfo.is_xrefed is False:
+            self.bdfInfo.cross_reference()
+            self.bdfInfo.is_xrefed = True
+
+        # Create a dictionary to sort all elements by property number
+        elemDict = {}
+        for elementID in self.bdfInfo.elements:
+            element = self.bdfInfo.elements[elementID]
+            propertyID = element.pid
+            if propertyID not in elemDict:
+                elemDict[propertyID] = {}
+                elemDict[propertyID]['elements'] = []
+                elemDict[propertyID]['dvs'] = {}
+            elemDict[propertyID]['elements'].append(element)
+
+        # Create a dictionary to sort all design variables
+        for dv in self.bdfInfo.dvprels:
+            propertyID = self.bdfInfo.dvprels[dv].pid
+            dvName = self.bdfInfo.dvprels[dv].pname_fid
+            self.dvNum = max(self.dvNum, self.bdfInfo.dvprels[dv].dvids[0])
+            elemDict[propertyID]['dvs'][dvName] = self.bdfInfo.dvprels[dv]
+        # Create option for user to specify scale values in BDF
+        self.scaleList = [1.0] * self.dvNum
+
+        # Callback function to return appropriate tacs MaterialProperties object
+        # For a pynastran mat card
+        def matCallBack(matInfo):
+            # First we define the material property object
+            if matInfo.type == 'MAT1':
+                mat = tacs.constitutive.MaterialProperties(rho=matInfo.rho, E=matInfo.e,
+                                                           nu=matInfo.nu, ys=matInfo.St,
+                                                           alpha=matInfo.a)
+            elif matInfo.type == 'MAT8':
+                E1 = matInfo.e11
+                E2 = matInfo.e22
+                nu12 = matInfo.nu12
+                G12 = matInfo.g12
+                G13 = matInfo.g1z
+                G23 = matInfo.g2z
+                # If out-of-plane shear values are 0, Nastran defaults them to the in-plane
+                if G13 == 0.0:
+                    G13 = G12
+                if G23 == 0.0:
+                    G23 = G12
+                rho = matInfo.rho
+                Xt = matInfo.Xt
+                Xc = matInfo.Xc
+                Yt = matInfo.Yt
+                Yc = matInfo.Yc
+                S12 = matInfo.S
+                # TODO: add alpha
+                mat = tacs.constitutive.MaterialProperties(rho=rho, E1=E1, E2=E2, nu12=nu12, G12=G12, G13=G13, G23=G23,
+                                                           Xt=Xt, Xc=Xc, Yt=Yt, Yc=Yc, S12=S12)
+            else:
+                raise Error("Unsupported material type '%s' for material number %d. " % (matInfo.type, matInfo.mid))
+
+            return mat
+
+        def elemCallBack(dvNum, compID, compDescript, elemDescripts, globalDVs, **kwargs):
+            # Initialize scale list for design variables we will add
+            scaleList = []
+
+            # Get the Nastran property ID
+            propertyID = kwargs['propID']
+            propInfo = self.bdfInfo.properties[propertyID]
+            elemInfo = elemDict[propertyID]['elements'][0]
+
+            # First we define the material object
+            # This property only references one material
+            if hasattr(propInfo, 'mid_ref'):
+                matInfo = propInfo.mid_ref
+                mat = matCallBack(matInfo)
+            # This property references multiple materials (maybe a laminate)
+            elif hasattr(propInfo, 'mids_ref'):
+                mat = []
+                for matInfo in propInfo.mids_ref:
+                    mat.append(matCallBack(matInfo))
+
+            # Next we define the constitutive object
+            if propInfo.type == 'PSHELL':  # Nastran isotropic shell
+                kcorr = propInfo.tst
+
+                if 'T' in elemDict[propertyID]['dvs']:
+                    thickness = elemDict[propertyID]['dvs']['T'].dvids_ref[0].xinit
+                    tNum = elemDict[propertyID]['dvs']['T'].dvids[0] - 1
+                    minThickness = elemDict[propertyID]['dvs']['T'].dvids_ref[0].xlb
+                    maxThickness = elemDict[propertyID]['dvs']['T'].dvids_ref[0].xub
+                    name = elemDict[propertyID]['dvs']['T'].dvids_ref[0].label
+                    self.scaleList[tNum - 1] = elemDict[propertyID]['dvs']['T'].coeffs[0]
+                else:
+                    thickness = propInfo.t
+                    tNum = -1
+                    minThickness = 0.0
+                    maxThickness = 1e20
+
+                con = tacs.constitutive.IsoShellConstitutive(mat, t=thickness,
+                                                             tlb=minThickness, tub=maxThickness, tNum=tNum)
+
+            elif propInfo.type == 'PCOMP':  # Nastran composite shell
+                numPlies = propInfo.nplies
+                plyThicknesses = []
+                plyAngles = []
+                plyMats = []
+
+                # if the laminate is symmetric, mirror the ply indices
+                if propInfo.lam == 'SYM':
+                    plyIndices = list(range(numPlies / 2))
+                    plyIndices.extend(plyIndices[::-1])
+                else:
+                    plyIndices = range(numPlies)
+
+                # Loop through plies and setup each entry in layup
+                for ply_i in plyIndices:
+                    plyThicknesses.append(propInfo.thicknesses[ply_i])
+                    plyMat = tacs.constitutive.OrthotropicPly(plyThicknesses[ply_i], mat[ply_i])
+                    plyMats.append(plyMat)
+                    plyAngles.append(propInfo.thetas[ply_i] * DEG2RAD)
+
+                # Convert thickness/angles to appropriate numpy array
+                plyThicknesses = np.array(plyThicknesses, dtype=self.dtype)
+                plyAngles = np.array(plyAngles, dtype=self.dtype)
+
+                if propInfo.lam is None or propInfo.lam in ['SYM', 'MEM']:
+                    # Discrete laminate class (not for optimization)
+                    con = tacs.constitutive.CompositeShellConstitutive(plyMats, plyThicknesses, plyAngles)
+                    # Need to add functionality to consider only membrane in TACS for type = MEM
+
+                else:
+                    raise Error("Unrecognized LAM type '%s' for PCOMP number %d." % (propInfo.lam, propertyID))
+
+            elif propInfo.type == 'PSOLID':  # Nastran solid property
+                if 'T' in elemDict[propertyID]['dvs']:
+                    thickness = elemDict[propertyID]['dvs']['T'].dvids_ref[0].xinit
+                    tNum = elemDict[propertyID]['dvs']['T'].dvids[0] - 1
+                    minThickness = elemDict[propertyID]['dvs']['T'].dvids_ref[0].xlb
+                    maxThickness = elemDict[propertyID]['dvs']['T'].dvids_ref[0].xub
+                    name = elemDict[propertyID]['dvs']['T'].dvids_ref[0].label
+                    self.scaleList[tNum - 1] = elemDict[propertyID]['dvs']['T'].coeffs[0]
+                else:
+                    thickness = 1.0
+                    tNum = -1
+                    minThickness = 0.0
+                    maxThickness = 10.0
+
+                con = tacs.constitutive.SolidConstitutive(mat, t=thickness,
+                                                          tlb=minThickness, tub=maxThickness, tNum=tNum)
+
+            else:
+                raise Error("Unsupported property type '%s' for property number %d. " % (propInfo.type, propertyID))
+
+            # Set up transform object which may be required for certain elements
+            transform = None
+            if hasattr(elemInfo, 'theta_mcid_ref'):
+                mcid = elemDict[propertyID]['elements'][0].theta_mcid_ref
+                if mcid:
+                    if mcid.type == 'CORD2R':
+                        refAxis = mcid.i
+                        transform = tacs.elements.ShellRefAxisTransform(refAxis)
+                    else:  # Don't support spherical/cylindrical yet
+                        raise Error("Unsupported material coordinate system type "
+                                    "'%s' for property number %d." % (mcid.type, propertyID))
+
+            # Finally set up the element objects belonging to this component
+            elemList = []
+            for descript in elemDescripts:
+                if descript in ['CQUAD4', 'CQUADR']:
+                    elem = tacs.elements.Quad4Shell(transform, con)
+                elif descript in ['CQUAD9', 'CQUAD']:
+                    elem = tacs.elements.Quad9Shell(transform, con)
+                elif descript in ['CTRIA3', 'CTRIAR']:
+                    elem = tacs.elements.Tri3Shell(transform, con)
+                elif 'CTETRA' in descript:
+                    # May have variable number of nodes in card
+                    nnodes = len(elemInfo.nodes)
+                    if nnodes == 4:
+                        basis = tacs.elements.LinearTetrahedralBasis()
+                    elif nnodes == 10:
+                        basis = tacs.elements.QuadraticTetrahedralBasis()
+                    else:
+                        raise Error("TACS does not currently support CTETRA elements with %d nodes." % nnodes)
+                    model = tacs.elements.LinearElasticity3D(con)
+                    elem = tacs.elements.Element3D(model, basis)
+                elif descript in ['CHEXA8', 'CHEXA']:
+                    basis = tacs.elements.LinearHexaBasis()
+                    model = tacs.elements.LinearElasticity3D(con)
+                    elem = tacs.elements.Element3D(model, basis)
+                else:
+                    raise Error("Unsupported element type "
+                                "'%s' specified for property number %d." % (descript, propertyID))
+                elemList.append(elem)
+
+            return elemList, scaleList
+
+        return elemCallBack
 
     ####### Static load methods ########
 
@@ -1483,7 +934,7 @@ class pyTACS(object):
                 allNodes = []
                 compIDs = set(compIDs)
                 for cID in compIDs:
-                    tmp = self.getConnectivityForComp(cID, nastranOrdering=True)
+                    tmp = self.meshLoader.getConnectivityForComp(cID, nastranOrdering=True)
                     allNodes.extend(self._flatten(tmp))
 
                 # Now just unique all the nodes:
@@ -1631,7 +1082,7 @@ class pyTACS(object):
                         "but length of vector provided was {}".format(vpn, len(F[0])))
 
         # First find the cooresponding local node ID on each processor
-        localNodeIDs = self.getLocalNodeIDsFromGlobal(nodeIDs, nastranOrdering)
+        localNodeIDs = self.meshLoader.getLocalNodeIDsFromGlobal(nodeIDs, nastranOrdering)
 
         # Set the structural problem
         self.setStructProblem(structProblem)
@@ -1684,7 +1135,7 @@ class pyTACS(object):
         tractions = np.atleast_1d(tractions)
 
         # Get global element IDs for the elements we're applying tractions to
-        elemIDs = self.getGlobalElementIDsForComps(compIDs, nastranOrdering=False)
+        elemIDs = self.meshLoader.getGlobalElementIDsForComps(compIDs, nastranOrdering=False)
         # Add tractions element by element
         self.addTractionToElements(structProblem, elemIDs, tractions, faceIndex, nastranOrdering=False)
 
@@ -1705,8 +1156,8 @@ class pyTACS(object):
         Parameters
         ----------
 
-        elemIDs : The components with added loads. Use selectCompIDs()
-            to determine this.
+        elemIDs : List
+            The global element ID numbers for which to apply the traction.
         tractions : Numpy 1d or 2d array length varsPerNodes or (elemIDs, varsPerNodes)
             Array of traction vectors for each element
         faceIndex : int
@@ -1716,17 +1167,12 @@ class pyTACS(object):
             Flag signaling whether elemIDs are in TACS (default)
             or NASTRAN ordering
         """
-        # Convert to tacs ordering for rest of procedure
-        if nastranOrdering:
-            globalElemIDs = self.idMap(elemIDs, self.nastranToTACSElemIDDict)
-        else:
-            globalElemIDs = elemIDs
 
         # Make sure the inputs are the correct shape
-        globalElemIDs = numpy.atleast_1d(globalElemIDs)
+        elemIDs = numpy.atleast_1d(elemIDs)
         tractions = numpy.atleast_2d(tractions).astype(dtype=self.dtype)
 
-        numElems = len(globalElemIDs)
+        numElems = len(elemIDs)
 
         # If the user only specified one traction vector,
         # we assume the force should be the same for each element
@@ -1738,7 +1184,7 @@ class pyTACS(object):
                         " {} tractions were specified for {} element IDs".format(tractions.shape[0], numElems))
 
         # First find the coresponding local element ID on each processor
-        localElemIDs = self.getLocalElementIDsFromGlobal(globalElemIDs, nastranOrdering=False)
+        localElemIDs = self.meshLoader.getLocalElementIDsFromGlobal(elemIDs, nastranOrdering=nastranOrdering)
 
         # Set the structural problem
         self.setStructProblem(structProblem)
@@ -1753,8 +1199,7 @@ class pyTACS(object):
                 # Mark element as found
                 elemFound[i] = 1
                 # Get the pointer for the tacs element object for this element
-                elemObjNum = self.elemObjectNumByElem[globalElemIDs[i]]
-                elemObj = self.elemObjects[elemObjNum]
+                elemObj = self.meshLoader.getElementObjectForElemID(elemIDs[i], nastranOrdering=nastranOrdering)
                 # Create appropriate traction object for this element type
                 tracObj = elemObj.createElementTraction(faceIndex, tractions[i])
                 # Traction not implemented for element
@@ -1805,7 +1250,7 @@ class pyTACS(object):
         pressures = np.atleast_1d(pressures)
 
         # Get global element IDs for the elements we're applying pressure to
-        elemIDs = self.getGlobalElementIDsForComps(compIDs, nastranOrdering=False)
+        elemIDs = self.meshLoader.getGlobalElementIDsForComps(compIDs, nastranOrdering=False)
         # Add pressure element by element
         self.addPressureToElements(structProblem, elemIDs, pressures, faceIndex, nastranOrdering=False)
 
@@ -1826,8 +1271,8 @@ class pyTACS(object):
         Parameters
         ----------
 
-        elemIDs : The components with added loads. Use selectCompIDs()
-            to determine this.
+        elemIDs : List
+            The global element ID numbers for which to apply the pressure.
         pressures : Numpy array length 1 or elemIDs
             Array of pressure values for each element
         faceIndex : int
@@ -1837,17 +1282,12 @@ class pyTACS(object):
             Flag signaling whether elemIDs are in TACS (default)
             or NASTRAN ordering
         """
-        # Convert to tacs ordering for rest of procedure
-        if nastranOrdering:
-            globalElemIDs = self.idMap(elemIDs, self.nastranToTACSElemIDDict)
-        else:
-            globalElemIDs = elemIDs
 
         # Make sure the inputs are the correct shape
-        globalElemIDs = numpy.atleast_1d(globalElemIDs)
+        elemIDs = numpy.atleast_1d(elemIDs)
         pressures = numpy.atleast_1d(pressures)
 
-        numElems = len(globalElemIDs)
+        numElems = len(elemIDs)
 
         # If the user only specified one pressure,
         # we assume the force should be the same for each element
@@ -1859,7 +1299,7 @@ class pyTACS(object):
                         " {} pressures were specified for {} element IDs".format(pressures.shape[0], numElems))
 
         # First find the coresponding local element ID on each processor
-        localElemIDs = self.getLocalElementIDsFromGlobal(globalElemIDs, nastranOrdering=False)
+        localElemIDs = self.meshLoader.getLocalElementIDsFromGlobal(elemIDs, nastranOrdering=nastranOrdering)
 
         # Set the structural problem
         self.setStructProblem(structProblem)
@@ -1873,11 +1313,7 @@ class pyTACS(object):
             if elemID >= 0:
                 elemFound[i] = 1
                 # Get the pointer for the tacs element object for this element
-                elemObjNum = self.elemObjectNumByElem[globalElemIDs[i]]
-                elemObj = self.elemObjects[elemObjNum]
-                # Get the pointer for the tacs element object for this element
-                elemObjNum = self.elemObjectNumByElem[globalElemIDs[i]]
-                elemObj = self.elemObjects[elemObjNum]
+                elemObj = self.meshLoader.getElementObjectForElemID(elemIDs[i], nastranOrdering=nastranOrdering)
                 # Create appropriate pressure object for this element type
                 pressObj = elemObj.createElementPressure(faceIndex, pressures[i])
                 # Pressure not implemented for element
@@ -1910,9 +1346,14 @@ class pyTACS(object):
         Automatically define tacs problem class using information contained in BDF file.
         This function assumes all loads are specified in the BDF and allows users to
         skip setting loads in Python.
-        NOTE: Currently only supports FORCE, MOMENT, PLOAD2, and PLOAD4 cards.
+        NOTE: Currently only supports LOAD, FORCE, MOMENT, PLOAD2, and PLOAD4 cards.
         NOTE: Currently only supports staticProblem (SOL 101)
         """
+
+        if self.assembler is None:
+            raise Error("TACS assembler has not been created. "
+                        "Assembler must created first by running 'createTACSAssembler' method.")
+
         # Make sure cross-referencing is turned on in pynastran
         if self.bdfInfo.is_xrefed is False:
             self.bdfInfo.cross_reference()
@@ -1947,90 +1388,101 @@ class pyTACS(object):
 
             if 'LOAD' in subCase.params:
                 loadsID = subCase.params['LOAD'][0]
-                # Check if there are any loads in this set
-                if loadsID in loads:
-                    loadSet = loads[loadsID]
+                # Get loads and scalers for this load case ID
+                loadSet, loadScale, _ = self.bdfInfo.get_reduced_loads(loadsID)
+                # Loop through every load in set and add it to problem
+                for loadInfo, scale in zip(loadSet, loadScale):
+                    # Add any point force or moment cards
+                    if loadInfo.type == 'FORCE' or loadInfo.type == 'MOMENT':
+                        nodeID = loadInfo.node_ref.nid
 
-                    loadDict = {}
-                    for load in loadSet:
-                        # Record any point force or moment cards
-                        if load.type == 'FORCE' or load.type == 'MOMENT':
-                            nodeID = load.node_ref.nid
+                        loadArray = numpy.zeros(vpn)
+                        if loadInfo.type == 'FORCE' and vpn >= 3:
+                            loadArray[:3] += scale * loadInfo.scaled_vector
+                        elif loadInfo.type == 'MOMENT' and vpn >= 6:
+                            loadArray[3:6] += scale * loadInfo.scaled_vector
+                        self.addLoadToNodes(sp, nodeID, loadArray, nastranOrdering=True)
 
-                            if nodeID not in loadDict:
-                                loadDict[nodeID] = numpy.zeros(vpn)
+                    # Add any pressure loads
+                    # Pressure load card specific to shell elements
+                    elif loadInfo.type == 'PLOAD2':
+                        elemIDs = loadInfo.eids
+                        pressure = scale * loadInfo.pressure
+                        self.addPressureToElements(sp, elemIDs, pressure, nastranOrdering=True)
 
-                            if load.type == 'FORCE' and vpn >= 3:
-                                loadDict[nodeID][:3] += load.scaled_vector
-                            elif load.type == 'MOMENT' and vpn >= 6:
-                                loadDict[nodeID][3:6] += load.scaled_vector
+                    # Alternate more general pressure load type
+                    elif loadInfo.type == 'PLOAD4':
+                        self._addPressureFromPLOAD4(sp, loadInfo, scale)
 
-                        # Add any pressure loads
-                        # Pressure load card specific to shell elements
-                        elif load.type == 'PLOAD2':
-                            elemIDs = load.eids
-                            pressure = load.pressure
-                            self.addPressureToElements(sp, elemIDs, pressure, nastranOrdering=True)
+                    else:
+                        TACSWarning("Unsupported load type "
+                                    " '%s' specified for load set number %d, skipping load" %(loadInfo.type, loadInfo.sid),
+                                    self.comm)
 
-                        # Alternate more general pressure load type
-                        elif load.type == 'PLOAD4':
-                            # Dictionary mapping nastran element face indices to TACS equivilent numbering
-                            nastranToTACSFaceIDDict = {'CTETRA': {1: 1, 2: 3, 3: 2, 4: 0},
-                                                       'CHEXA': {1: 4, 2: 2, 3: 0, 4: 3, 5: 0, 6: 5}}
-
-                            # We don't support pressure variation across elements, for now just average it
-                            pressure = np.mean(load.pressures)
-                            for elemInfo in load.eids_ref:
-                                elemID = elemInfo.eid
-
-                                # Get the correct face index number based on element type
-                                if 'CTETRA' in elemInfo.type:
-                                    for faceIndex in elemInfo.faces:
-                                        if load.g1 in elemInfo.faces[faceIndex] and \
-                                            load.g34 not in elemInfo.faces[faceIndex]:
-                                            faceIndex = nastranToTACSFaceIDDict['CTETRA'][faceIndex]
-                                            # Pressure orientation is flipped for solid elements per Nastran convention
-                                            pressure *= -1.0
-                                            break
-
-                                elif 'CHEXA' in elemInfo.type:
-                                    for faceIndex in elemInfo.faces:
-                                        if load.g1 in elemInfo.faces[faceIndex] and \
-                                            load.g34 in elemInfo.faces[faceIndex]:
-                                            faceIndex = nastranToTACSFaceIDDict['CHEXA'][faceIndex]
-                                            # Pressure orientation is flipped for solid elements per Nastran convention
-                                            pressure *= -1.0
-                                            break
-
-                                elif 'CQUAD' in elemInfo.type or 'CTRIA' in elemInfo.type:
-                                    # Face index doesn't matter for shells, just use 0
-                                    faceIndex = 0
-
-                                else:
-                                    raise Error("Unsupported element type "
-                                                "'%s' specified for PLOAD4 load set number %d." % (elemInfo.type, load.sid))
-
-                                # Figure out whether or not this is a traction based on if a vector is defined
-                                if np.linalg.norm(load.nvector) == 0.0:
-                                    self.addPressureToElements(sp, elemID, pressure, faceIndex,
-                                                               nastranOrdering=True)
-                                else:
-                                    trac = pressure * load.nvector
-                                    self.addTractionToElements(sp, elemID, trac, faceIndex,
-                                                               nastranOrdering=True)
-
-                        else:
-                            TACSWarning("Unsupported load type "
-                                        " '%s' specified for load set number %d, skipping load" %(load.type, load.sid),
-                                        self.comm)
-
-                    # Add all point forces/moments found for this problem at once
-                    if len(loadDict) > 0:
-                        self.addLoadToNodes(sp, list(loadDict.keys()), list(loadDict.values()), nastranOrdering=True)
             # append to list of structural problems
             structProblems[subCase.id] = sp
 
         return structProblems
+
+    def _addPressureFromPLOAD4(self, staticProb, loadInfo, scale=1.0):
+        """
+        Add pressure to tacs static problem from pynastran PLOAD4 card.
+        Should only be called by createTACSProbsFromBDF and not directly by user.
+        """
+        # Dictionary mapping nastran element face indices to TACS equivilent numbering
+        nastranToTACSFaceIDDict = {'CTETRA4': {1: 1, 2: 3, 3: 2, 4: 0},
+                                   'CTETRA': {2: 1, 4: 3, 3: 2, 1: 0},
+                                   'CHEXA': {1: 4, 2: 2, 3: 0, 4: 3, 5: 0, 6: 5}}
+
+        # We don't support pressure variation across elements, for now just average it
+        pressure = scale * np.mean(loadInfo.pressures)
+        for elemInfo in loadInfo.eids_ref:
+            elemID = elemInfo.eid
+
+            # Get the correct face index number based on element type
+            if 'CTETRA' in elemInfo.type:
+                for faceIndex in elemInfo.faces:
+                    if loadInfo.g1 in elemInfo.faces[faceIndex] and \
+                            loadInfo.g34 not in elemInfo.faces[faceIndex]:
+                        # For some reason CTETRA4 is the only element that doesn't
+                        # use ANSYS face numbering convention by default
+                        if len(elemInfo.nodes) == 4:
+                            faceIndex = nastranToTACSFaceIDDict['CTETRA4'][faceIndex]
+                        else:
+                            faceIndex = nastranToTACSFaceIDDict['CTETRA'][faceIndex]
+                        # Positive pressure is inward for solid elements, flip pressure if necessary
+                        # We don't flip it for face 0, because the normal for that face points inward by convention
+                        # while the rest point outward
+                        if faceIndex != 0:
+                            pressure *= -1.0
+                        break
+
+            elif 'CHEXA' in elemInfo.type:
+                for faceIndex in elemInfo.faces:
+                    if loadInfo.g1 in elemInfo.faces[faceIndex] and \
+                            loadInfo.g34 in elemInfo.faces[faceIndex]:
+                        faceIndex = nastranToTACSFaceIDDict['CHEXA'][faceIndex]
+                        # Pressure orientation is flipped for solid elements per Nastran convention
+                        pressure *= -1.0
+                        break
+
+            elif 'CQUAD' in elemInfo.type or 'CTRIA' in elemInfo.type:
+                # Face index doesn't matter for shells, just use 0
+                faceIndex = 0
+
+            else:
+                raise Error("Unsupported element type "
+                            "'%s' specified for PLOAD4 load set number %d." % (elemInfo.type, loadInfo.sid))
+
+            # Figure out whether or not this is a traction based on if a vector is defined
+            if np.linalg.norm(loadInfo.nvector) == 0.0:
+                self.addPressureToElements(staticProb, elemID, pressure, faceIndex,
+                                           nastranOrdering=True)
+            else:
+                trac = pressure * loadInfo.nvector
+                self.addTractionToElements(staticProb, elemID, trac, faceIndex,
+                                           nastranOrdering=True)
+
 
     ####### Static solver methods ########
 
@@ -2822,7 +2274,7 @@ class pyTACS(object):
 
         self.fam = []
         for i in range(self.nComp):
-            aux = self.compDescript[i].split(self.getOption('familySeparator'))
+            aux = self.compDescripts[i].split(self.getOption('familySeparator'))
             self.fam.append(aux[0])
 
         # Uniqify them and sort
@@ -2830,7 +2282,7 @@ class pyTACS(object):
 
         self.compFam = numpy.zeros(self.nComp, dtype='intc')
         for i in range(self.nComp):
-            aux = self.compDescript[i].split(self.getOption('familySeparator'))
+            aux = self.compDescripts[i].split(self.getOption('familySeparator'))
             self.compFam[i] = self.fam.index(aux[0])
 
     def _createOutputViewer(self):
@@ -2896,11 +2348,11 @@ class pyTACS(object):
 
             elif isinstance(item, str):
                 # This is a little inefficinet here; loop over
-                # self.compDescript and see if 'item' (a string) in
+                # self.compDescripts and see if 'item' (a string) in
                 # part of the description. if so add.
                 item = item.upper()
                 for i in range(self.nComp):
-                    if item in self.compDescript[i]:
+                    if item in self.compDescripts[i]:
                         compIDs[-1].append(i)
             else:
                 TACSWarning('Unidentifiable information given for \'include\'\
@@ -2925,123 +2377,6 @@ class pyTACS(object):
 
         return compIDs
 
-    def _getCompIDsInBounds(self, includeBounds, projectVector=None):
-
-        # This function Requires scipy spatital for doing the convex
-        # hull in (either 2d or 3d)
-
-        if self.rank == 0:
-            try:
-                from scipy.spatial import Delaunay
-            except ImportError:
-                raise Error("scipy.spatial could not be loaded. Selecting "
-                            "components by bounds is not supported.")
-
-            # Use default projectVector if not given:
-            if projectVector is None:
-                projectVector = numpy.array(self.getOption('projectVector'),
-                                            dtype='d')
-            else:
-                projectVector = numpy.array(projectVector, dtype='d')
-
-            # Individually select the compIDs for each entry in
-            # includeBounds...this is recursive.
-            compIDs = {}
-            for item in includeBounds:
-                compIDs[item] = self.selectCompIDs(include=item)[0]
-
-            # Now we want to find the "intersection" between each of the
-            # sets of componets in compIDs. I don't think there is a
-            # better way than an n^2 algorithm for this.
-
-            shared_edges = []
-            keys = sorted(compIDs.keys())
-            for i in range(len(keys)):  # |-> Double loop over each
-                for j in range(i + 1, len(keys)):  # |   group of items
-                    if i != j:
-                        iComps = compIDs[keys[i]]  # | -> These are lists of the
-                        jComps = compIDs[keys[j]]  # | -> compontIDs
-
-                        # Now we want to find the edges shared between
-                        # each possible set of components
-                        for ii in iComps:  # | -> Double group over each
-                            for jj in jComps:  # | pair of componets
-                                if ii < jj:
-                                    key = (ii, jj)
-                                else:
-                                    key = (jj, ii)
-
-                                if key in self.dvEdges:
-                                    # These are the shared edges.
-                                    shared_edges.append(self.dvEdges[key])
-
-            # That finishes the logical section. Now we proceed
-            # geometrically. Get a unique set of nodes:
-            uniqueNodes = numpy.array(numpy.unique(self._flatten(shared_edges)),
-                                      dtype='intc')
-            # Extract the coordinates for these
-            nodes = self.getBDFNodes(uniqueNodes)
-
-            # We first do a 3d bounding box (+ delta) elimination of
-            # possible components. This is by far the fastest way to
-            # eliminate components:
-            xMin = numpy.min(nodes, axis=0)
-            xMax = numpy.max(nodes, axis=0)
-            # Expand the box by 10%
-            d = numpy.linalg.norm(xMax - xMin) * 0.1
-            xMin -= d
-            xMax += d
-
-            # Now we have to contruct a coordinate system in the plane
-            # defined by projectVector.
-            length = numpy.linalg.norm(projectVector)
-            if length < 1e-12:
-                raise Error("projectVector has almost zero length!")
-
-            n = projectVector / length  # Plane (unit) normal
-            pt = [1, 1, 1]
-            if numpy.linalg.norm(numpy.cross(pt, n)) < 1e-12:
-                pt = [2, 1, 1]
-
-            v1 = pt - numpy.dot(pt, n) * n
-            v1 /= numpy.linalg.norm(v1)
-            v2 = numpy.cross(v1, projectVector)
-
-            def projectNodesOntoPlane(nodes, projectVector):
-                projNodes = numpy.zeros((len(nodes), 2))
-                for i in range(len(nodes)):
-                    # Project onto v1
-                    projNodes[i, 0] = numpy.dot(nodes[i], v1)
-                    projNodes[i, 1] = numpy.dot(nodes[i] -
-                                                projNodes[i, 0] * v1, v2)
-                return projNodes
-
-            # Project to plane
-            nodes = projectNodesOntoPlane(nodes, projectVector)
-
-            # Create delaunay triagularization
-            hull = Delaunay(nodes)
-
-            # Simple loop over each component:
-            compIDs = []
-            for compID in range(self.nComp):
-                cenPt = self.compCenters[compID]
-                if (cenPt[0] > xMin[0] and cenPt[0] < xMax[0] and
-                        cenPt[1] > xMin[1] and cenPt[1] < xMax[1] and
-                        cenPt[2] > xMin[2] and cenPt[2] < xMax[2]):
-                    # We are in the bounds, so check in hull:
-                    proj = projectNodesOntoPlane([cenPt],
-                                                 projectVector)[0]
-                    res = hull.find_simplex(proj)
-                    if res >= 0:
-                        compIDs.append(compID)
-
-        else:  # rank == 0
-            compIDs = None
-
-        # Broadcast and return
-        return self.comm.bcast(compIDs, root=0)
-
     def _createElements(self, elemCallBack):
         """
         Create all the constitutive objects by calling the
@@ -3051,7 +2386,7 @@ class pyTACS(object):
         for i in range(self.nComp):
 
             # Get a list of compDescripts to help the user
-            compDescript = self.compDescript[i]
+            compDescript = self.compDescripts[i]
             numElements = len(self.elemDescripts[i])
             # TACS component ID
             compID = i
@@ -3173,8 +2508,7 @@ class pyTACS(object):
             # there may be multiple (e.g CQUAD4 + CTRIA3)
             for j, elemObject in enumerate(elemObjects):
                 # Set each of the elements for this component
-                pointer = self.elemObjectNumByComp[i][j]
-                self.elemObjects[pointer] = elemObject
+                self.meshLoader.setElementObject(i, j, elemObject)
                 # set varsPerNode
                 elemVarsPerNode = elemObject.getVarsPerNode()
                 if self.varsPerNode is None:
@@ -3454,40 +2788,6 @@ class pyTACS(object):
         assert len(field) == 8, ('value=|%s| field=|%s| is not 8 characters '
                                  'long, its %s' % (value, field, len(field)))
         return field
-
-    def isDOFInString(self, dofString, numDOFs):
-        """
-        Converts a dof string to a boolean list.
-        Examples:
-            '123' -> [1, 1, 1, 0, 0, 0]
-            '1346' -> [1, 0, 1, 1, 0, 1]
-        """
-        dofList = []
-        for dof in range(numDOFs):
-            dof = str(dof + 1)
-            loc = dofString.find(dof)
-            if loc > -1:
-                dofList.append(1)
-            else:
-                dofList.append(0)
-        return dofList
-
-    def idMap(self, fromIDs, tacsIDDict):
-        """
-        Translate fromIDs numbering from nastran numbering to tacs numbering.
-        If node ID doesn't exist in nastranIDList, return -1 for entry
-        """
-        if isinstance(fromIDs, int):
-            return tacsIDDict[fromIDs]
-
-        toIDs = [None] * len(fromIDs)
-        for i, id in enumerate(fromIDs):
-            if id in tacsIDDict:
-                toIDs[i] = tacsIDDict[id]
-            else:
-                toIDs[i] = -1
-
-        return toIDs
 
 
 class TACSLoadCase(object):
