+++ tests/integration_tests/test_elast_beam_quadtetra_element_3d.py
@@ -4,21 +4,20 @@ from tacs import pytacs, TACS, elements, constitutive, functions, problems
 from pytacs_analysis_base_test import PyTACSTestCase
 
 '''
-Cylindrical beam constructed from shell elements. The beam is cantilevered at
-one end and loaded at the other using an RBE3.
+ 10m x 1m x 1m Square beam constructed from tetrahedral elements. The beam is cantilevered at
+one end and loaded using a pressure on the opposite face. This leads to a pure axial loading on the model.
 
 test StructuralMass and Compliance functions and sensitivities
 '''
 
 base_dir = os.path.dirname(os.path.abspath(__file__))
-bdf_file = os.path.join(base_dir, "./input_files/cylinder.bdf")
+bdf_file = os.path.join(base_dir, "./input_files/solid_beam.bdf")
 
-FUNC_REFS = {'Axial_compliance': 0.004606292990644062, 'Axial_mass': 208.58647120309135,
-             'Shear-Bending_compliance': 0.32883253964394005, 'Shear-Bending_mass': 208.58647120309135,
-             'Moment-Bending_compliance': 0.03744304905762492, 'Moment-Bending_mass': 208.58647120309135,
-             'Torsion_compliance': 0.05100479047600847, 'Torsion_mass': 208.58647120309135,
-             'Combined_compliance': 0.4218866721681868, 'Combined_mass': 208.58647120309135}
+FUNC_REFS = {'AxialPressure_compliance': 14226.813534942703,
+             'AxialPressure_ks_disp': 0.3088436763967383,
+             'AxialPressure_mass': 27000.00000000002}
 
+ksweight = 10.0
 
 class ProblemTest(PyTACSTestCase.PyTACSTest):
     N_PROCS = 2  # this is how many MPI processes to use for this TestCase.
@@ -33,8 +32,8 @@ class ProblemTest(PyTACSTestCase.PyTACSTest):
             self.atol = 1e-8
             self.dh = 1e-50
         else:
-            self.rtol = 2e-1
-            self.atol = 1e-3
+            self.rtol = 1e-6
+            self.atol = 1e-2
             self.dh = 1e-6
 
         # Instantiate FEA Solver
@@ -67,7 +66,9 @@ class ProblemTest(PyTACSTestCase.PyTACSTest):
         # Add Functions
         fea_solver.addFunction('mass', functions.StructuralMass)
         fea_solver.addFunction('compliance', functions.Compliance)
-        func_list = ['mass', 'compliance']
+        fea_solver.addFunction('ks_disp', functions.KSDisplacement,
+                               ksWeight=ksweight, direction=[-100.0, -100.0, -100.0])
+        func_list = ['mass', 'compliance', 'ks_disp']
         return func_list, FUNC_REFS
 
     def setup_tacs_problems(self, fea_solver):
